SA-19-Move  HUGGING_FACE request to server side to do not expose the authorization token
